fix: reorganize code and move test
